Refactor 소셜 로그인 테스트용 컨트롤러 구분
